# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id?: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string | null;
  role?: string;
  isAuthenticated: boolean;
}

export interface User {
  id: string;
  username: string;
  firstName: string;
  lastName: string;
  profileImage?: string | null;
  role: string;
  department?: string | null;
  title?: string | null;
  createdAt: string;
}

export interface CreateUserBody {
  firstName: string;
  lastName: string;
  role?: string;
  department?: string | null;
  title?: string | null;
}

export interface UpdateUserBody {
  firstName?: string;
  lastName?: string;
  role?: string;
  department?: string | null;
  title?: string | null;
}

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: string;
  priority: string;
  assigneeId?: string | null;
  assignee?: User | null;
  createdById: string;
  dueDate?: string | null;
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  status?: string;
  priority?: string;
  assigneeId?: string | null;
  dueDate?: string | null;
}

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  status?: string;
  priority?: string;
  assigneeId?: string | null;
  dueDate?: string | null;
  completedAt?: string | null;
}

export interface Kpi {
  id: number;
  name: string;
  description?: string | null;
  unit: string;
  targetValue: number;
  userId: string;
  period: string;
  createdAt: string;
}

export interface CreateKpiBody {
  name: string;
  description?: string | null;
  unit: string;
  targetValue: number;
  userId: string;
  period: string;
}

export interface UpdateKpiBody {
  name?: string;
  description?: string | null;
  unit?: string;
  targetValue?: number;
  period?: string;
}

export interface KpiEntry {
  id: number;
  kpiId: number;
  kpi?: Kpi | null;
  userId: string;
  actualValue: number;
  notes?: string | null;
  recordedAt: string;
  createdAt: string;
}

export interface CreateKpiEntryBody {
  kpiId: number;
  userId: string;
  actualValue: number;
  notes?: string | null;
  recordedAt: string;
}

export interface AttendanceRecord {
  id: number;
  userId: string;
  user?: User | null;
  clockIn: string;
  clockOut?: string | null;
  date: string;
  status: string;
  notes?: string | null;
  createdAt: string;
}

export interface QualityCheck {
  id: number;
  taskId?: number | null;
  reviewerId: string;
  reviewer?: User | null;
  submitterId: string;
  submitter?: User | null;
  rating: number;
  feedback?: string | null;
  status: string;
  videoVersionId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateQualityCheckBody {
  taskId?: number | null;
  submitterId: string;
  rating: number;
  feedback?: string | null;
  status?: string;
  videoVersionId?: number | null;
}

export interface UpdateQualityCheckBody {
  rating?: number;
  feedback?: string | null;
  status?: string;
}

export interface Message {
  id: number;
  content: string;
  authorId: string;
  author?: User | null;
  parentId?: number | null;
  createdAt: string;
}

export interface CreateMessageBody {
  content: string;
  parentId?: number | null;
}

export interface Meeting {
  id: number;
  title: string;
  description?: string | null;
  scheduledAt: string;
  duration: number;
  meetingUrl?: string | null;
  organizerId: string;
  organizer?: User | null;
  attendees?: User[];
  createdAt: string;
}

export interface CreateMeetingBody {
  title: string;
  description?: string | null;
  scheduledAt: string;
  duration: number;
  meetingUrl?: string | null;
  attendeeIds?: string[];
}

export interface UpdateMeetingBody {
  title?: string;
  description?: string | null;
  scheduledAt?: string;
  duration?: number;
  meetingUrl?: string | null;
  attendeeIds?: string[];
}

export interface Notification {
  id: number;
  userId: string;
  type: string;
  title: string;
  body?: string | null;
  isRead: boolean;
  linkUrl?: string | null;
  createdAt: string;
}

export interface LeaderboardEntry {
  userId: string;
  user?: User;
  rank: number;
  score: number;
  kpiScore: number;
  attendanceScore: number;
  qualityScore: number;
  onTimeScore: number;
  month: number;
  year: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface VideoProject {
  id: number;
  title: string;
  clientName?: string | null;
  description?: string | null;
  taskId?: number | null;
  status: string;
  createdById: string;
  latestVersionId?: number | null;
  latestVersionNumber?: number | null;
  latestVersionStatus?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface VideoVersion {
  id: number;
  projectId: number;
  versionNumber: number;
  objectPath: string;
  fileName: string;
  fileSize: number;
  status: string;
  uploadedById: string;
  uploadedBy?: User | null;
  shareToken?: string | null;
  createdAt: string;
}

export interface VideoProjectDetail {
  id: number;
  title: string;
  clientName?: string | null;
  description?: string | null;
  taskId?: number | null;
  status: string;
  createdById: string;
  versions: VideoVersion[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateVideoProjectBody {
  title: string;
  clientName?: string | null;
  description?: string | null;
  taskId?: number | null;
}

export interface UpdateVideoProjectBody {
  title?: string;
  clientName?: string | null;
  description?: string | null;
  status?: string;
}

export interface CreateVideoVersionBody {
  objectPath: string;
  fileName: string;
  fileSize: number;
}

export interface VideoComment {
  id: number;
  versionId: number;
  authorId?: string | null;
  authorName: string;
  authorType: string;
  content: string;
  timestampSeconds?: number | null;
  isResolved: boolean;
  createdAt: string;
}

export interface CreateVideoCommentBody {
  content: string;
  timestampSeconds?: number | null;
}

export interface ShareToken {
  token: string;
  shareUrl: string;
  versionId: number;
  expiresAt?: string | null;
}

export interface PublicReviewData {
  projectTitle: string;
  clientName?: string | null;
  versionNumber: number;
  objectPath: string;
  fileName: string;
  comments: VideoComment[];
}

export interface CreatePublicCommentBody {
  content: string;
  authorName: string;
  timestampSeconds?: number | null;
}

export type ListTasksParams = {
  assigneeId?: string;
  status?: string;
};

export type ListKpiEntriesParams = {
  userId?: string;
  kpiId?: number;
};

export type ListAttendanceParams = {
  userId?: string;
  date?: string;
};

export type ListQualityChecksParams = {
  taskId?: number;
};

export type ListMessagesParams = {
  limit?: number;
  before?: number;
};

export type GetLeaderboardParams = {
  month?: number;
  year?: number;
};

export type ListVideoProjectsParams = {
  taskId?: number;
};