# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Setspace Agency Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthenticated
...
  /kpis:
    get:
      operationId: listKpis
      tags: [kpis]
      summary: List KPI definitions
      responses:
        "200":
          description: KPI list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
    post:
      operationId: createKpi
      tags: [kpis]
      summary: Create a KPI definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiBody"
      responses:
        "201":
          description: Created KPI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
  /kpis/{kpiId}:
    patch:
      operationId: updateKpi
      tags: [kpis]
      summary: Update a KPI
      parameters:
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKpiBody"
      responses:
        "200":
          description: Updated KPI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    delete:
      operationId: deleteKpi
      tags: [kpis]
      summary: Delete a KPI
      parameters:
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /kpi-entries:
    get:
      operationId: listKpiEntries
      tags: [kpis]
      summary: List KPI entries
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: kpiId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: KPI entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiEntry"
    post:
      operationId: createKpiEntry
      tags: [kpis]
      summary: Record a KPI entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiEntry"
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{meetingId}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    patch:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeetingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
        isAuthenticated:
          type: boolean
      required: [isAuthenticated]
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, username, firstName, lastName, role, createdAt]
...
    Kpi:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        unit:
          type: string
        targetValue:
          type: number
        userId:
          type: string
        period:
          type: string
        createdAt:
          type: string
      required: [id, name, unit, targetValue, userId, period, createdAt]
    CreateKpiBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        unit:
          type: string
        targetValue:
          type: number
        userId:
          type: string
        period:
          type: string
      required: [name, unit, targetValue, userId, period]
    UpdateKpiBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        unit:
          type: string
        targetValue:
          type: number
        period:
          type: string
    KpiEntry:
      type: object
      properties:
        id:
          type: integer
        kpiId:
          type: integer
        kpi:
          $ref: "#/components/schemas/Kpi"
          nullable: true
        userId:
          type: string
        actualValue:
          type: number
        notes:
          type: string
          nullable: true
        recordedAt:
          type: string
        createdAt:
          type: string
      required: [id, kpiId, userId, actualValue, recordedAt, createdAt]
    CreateKpiEntryBody:
      type: object
      properties:
        kpiId:
          type: integer
        userId:
          type: string
        actualValue:
          type: number
        notes:
          type: string
          nullable: true
        recordedAt:
          type: string
      required: [kpiId, userId, actualValue, recordedAt]
...
    Meeting:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        scheduledAt:
          type: string
        duration:
          type: integer
        meetingUrl:
          type: string
          nullable: true
        organizerId:
          type: string
        organizer:
          $ref: "#/components/schemas/User"
          nullable: true
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/User"
        createdAt:
          type: string
      required: [id, title, scheduledAt, duration, organizerId, createdAt]
    CreateMeetingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        scheduledAt:
          type: string
        duration:
          type: integer
        meetingUrl:
          type: string
          nullable: true
        attendeeIds:
          type: array
          items:
            type: string
      required: [title, scheduledAt, duration]
    UpdateMeetingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        scheduledAt:
          type: string
        duration:
          type: integer
        meetingUrl:
          type: string
          nullable: true
        attendeeIds:
          type: array
          items:
            type: string
...