# lib/api-zod/src/generated/types/kpiEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Kpi } from "./kpi";

export interface KpiEntry {
  id: number;
  kpiId: number;
  kpi?: Kpi | null;
  userId: string;
  actualValue: number;
  notes?: string | null;
  recordedAt: string;
  createdAt: string;
}