# lib/api-zod/src/generated/types/updateKpiBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateKpiBody {
  name?: string;
  description?: string | null;
  unit?: string;
  targetValue?: number;
  period?: string;
}