# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  firstName: string;
  lastName: string;
  profileImage?: string | null;
  role: string;
  department?: string | null;
  title?: string | null;
  createdAt: string;
}